# docs/swagger.yaml
basePath: /
definitions:
...
  todo_internal_models.CreateReminderConfigRequest:
    properties:
      channel_type:
        enum:
        - webhook
        - feishu
        - dingtalk
        - wecom
        - slack
        type: string
      max_retries:
        maximum: 10
        minimum: 0
        type: integer
      name:
        maxLength: 64
        type: string
      retry_delay_seconds:
        maximum: 300
        minimum: 1
        type: integer
      webhook_body_template:
        type: string
      webhook_headers:
        additionalProperties:
          type: string
        type: object
      webhook_method:
        enum:
        - GET
        - POST
        - PUT
        type: string
      webhook_url:
        type: string
    required:
    - channel_type
    - name
    - webhook_url
    type: object
...
  todo_internal_models.UpdateReminderConfigRequest:
    properties:
      channel_type:
        enum:
        - webhook
        - feishu
        - dingtalk
        - wecom
        - slack
        type: string
      enabled:
        type: boolean
      max_retries:
        maximum: 10
        minimum: 0
        type: integer
      name:
        maxLength: 64
        type: string
      retry_delay_seconds:
        maximum: 300
        minimum: 1
        type: integer
      webhook_body_template:
        type: string
      webhook_headers:
        additionalProperties:
          type: string
        type: object
      webhook_method:
        enum:
        - GET
        - POST
        - PUT
        type: string
      webhook_url:
        type: string
    type: object
...
  todo_internal_models.UserReminderConfig:
    properties:
      channel_type:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      max_retries:
        type: integer
      name:
        type: string
      retry_delay_seconds:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      webhook_body_template:
        type: string
      webhook_headers:
        additionalProperties:
          type: string
        type: object
      webhook_method:
        type: string
      webhook_url:
        type: string
    type: object
...
  todo_internal_utils.ErrorResponse:
    properties:
      code:
        example: NOT_FOUND
        type: string
      error:
        example: task not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  todo_internal_utils.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 多用户 TODO 任务管理系统，支持用户注册/登录、个人 API Key、多渠道提醒推送
  title: TODO 任务管理系统 API
  version: 2.0.0
paths:
...
  /api/v1/user/reminder-configs:
    get:
      description: 获取当前用户的所有通知渠道配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/todo_internal_models.UserReminderConfig'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 列出提醒配置
      tags:
      - reminder-config
    post:
      consumes:
      - application/json
      description: 为当前用户创建一个新的通知渠道配置
      parameters:
      - description: 配置信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.CreateReminderConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.UserReminderConfig'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建提醒配置
      tags:
      - reminder-config
  /api/v1/user/reminder-configs/{id}:
    delete:
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除提醒配置
      tags:
      - reminder-config
    get:
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.UserReminderConfig'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取单个提醒配置
      tags:
      - reminder-config
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.UpdateReminderConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.UserReminderConfig'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新提醒配置
      tags:
      - reminder-config
securityDefinitions:
  APIKeyAuth:
    in: header
    name: api-key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"